Add testing catalog
